+++ server/RdtClient.Data/Data/DataContext.cs
@@ -66,6 +66,12 @@ namespace RdtClient.Data.Data
                 },
                 new Setting
                 {
+                    SettingId = "TorrentLifetime",
+                    Type = "Int32",
+                    Value = "0"
+                },
+                new Setting
+                {
                     SettingId = "RealDebridApiKey",
                     Type = "String",
                     Value = ""