+++ server/RdtClient.Data/Models/Data/Torrent.cs
@@ -35,6 +35,7 @@ namespace RdtClient.Data.Models.Data
         public Int32 DownloadRetryAttempts { get; set; }
         public Int32 TorrentRetryAttempts { get; set; }
         public Int32 DeleteOnError { get; set; }
+        public Int32 Lifetime { get; set; }
 
         public String Error { get; set; }