+++ server/RdtClient.Data/Models/Internal/DbSettings.cs
@@ -26,5 +26,6 @@ namespace RdtClient.Data.Models.Internal
         public Int32 DownloadRetryAttempts { get; set; }
         public Int32 TorrentRetryAttempts { get; set; }
         public Int32 DeleteOnError { get; set; }
+        public Int32 TorrentLifetime { get; set; }
     }
 }